Localize warp query lookup to merge attention
