+++ vllm/v1/attention/ops/triton_attention_helpers.py
@@ -107,30 +107,6 @@ def find_seq_idx(
 
 
 @triton.jit
-def find_seq_idx_warp(
-    query_start_len_ptr,
-    target_idx,
-    num_seqs,
-    BLOCK_Q: tl.constexpr,
-):
-    """Resolve a ragged query block by scanning 32 boundaries at a time."""
-    lanes = tl.arange(0, 32)
-    chunk_start = 0
-    seq_idx = 0
-    move_on = True
-    while (chunk_start < num_seqs) & move_on:
-        seq_offsets = chunk_start + lanes
-        seq_mask = seq_offsets < num_seqs
-        query_starts = tl.load(query_start_len_ptr + seq_offsets, mask=seq_mask)
-        seq_block_starts = query_starts // BLOCK_Q + seq_offsets
-        match_count = tl.sum((seq_mask & (seq_block_starts <= target_idx)).to(tl.int32))
-        seq_idx = chunk_start + match_count - 1
-        move_on = match_count == 32
-        chunk_start += 32
-    return seq_idx
-
-
-@triton.jit
 def init_softmax_M(
     sink_ptr,
     query_offset_1,
